+++ crates/swc/tests/fixture/issues-3xxx/3126/1/output/index.js
@@ -1,15 +1,12 @@
 !function() {
-    var __webpack_modules__ = {
+    __webpack_require__.m = {
         746: function(__unused_webpack_module, __unused_webpack___webpack_exports__, __webpack_require__1) {
             Object.prototype.hasOwnProperty;
         }
-    };
-    __webpack_require__.m = __webpack_modules__, __webpack_require__.O = function(result, chunkIds, fn, priority) {
-        for(var _loop = function(j) {
-            Object.keys(__webpack_require__.O).every(function(key) {
-                return __webpack_require__.O[key](chunkIds[j]);
-            });
-        }, j = 0; j < chunkIds.length; j++)_loop(j);
+    }, __webpack_require__.O = function(result, chunkIds, fn, priority) {
+        for(var j = 0; j < chunkIds.length; j++)Object.keys(__webpack_require__.O).every(function(key) {
+            return __webpack_require__.O[key](chunkIds[j]);
+        });
     };
     var __webpack_exports__ = __webpack_require__.O(void 0, [
         532