+++ crates/swc/tests/tsc-references/optionalChainingInLoop_es5.1.normal.js
@@ -5,14 +5,14 @@
 var list = [];
 var _iteratorNormalCompletion = true, _didIteratorError = false, _iteratorError = undefined;
 try {
-    var _loop = function(_iterator, _step) {
+    var _loop = function() {
         var comp = _step.value;
         comp.sp.y = comp.sp.r.find(function(k) {
             return k.c == (comp.xp ? "1" : "0");
         });
         var _iteratorNormalCompletion = true, _didIteratorError = false, _iteratorError = undefined;
         try {
-            for(var _iterator1 = comp.c[Symbol.iterator](), _step1; !(_iteratorNormalCompletion = (_step1 = _iterator1.next()).done); _iteratorNormalCompletion = true){
+            for(var _iterator = comp.c[Symbol.iterator](), _step1; !(_iteratorNormalCompletion = (_step1 = _iterator.next()).done); _iteratorNormalCompletion = true){
                 var item = _step1.value;
                 var ref;
                 item.v = !!((ref = item.t) === null || ref === void 0 ? void 0 : ref.length);
@@ -22,8 +22,8 @@ try {
             _iteratorError = err;
         } finally{
             try {
-                if (!_iteratorNormalCompletion && _iterator1.return != null) {
-                    _iterator1.return();
+                if (!_iteratorNormalCompletion && _iterator.return != null) {
+                    _iterator.return();
                 }
             } finally{
                 if (_didIteratorError) {
@@ -32,7 +32,7 @@ try {
             }
         }
     };
-    for(var _iterator = list[Symbol.iterator](), _step; !(_iteratorNormalCompletion = (_step = _iterator.next()).done); _iteratorNormalCompletion = true)_loop(_iterator, _step);
+    for(var _iterator = list[Symbol.iterator](), _step; !(_iteratorNormalCompletion = (_step = _iterator.next()).done); _iteratorNormalCompletion = true)_loop();
 } catch (err) {
     _didIteratorError = true;
     _iteratorError = err;