+++ crates/swc_ecma_transforms_compat/tests/block-scoping/issue-5270/2/output.js
@@ -1,0 +1,3 @@
+for(var a = 0; a < 10; a++){
+    console.log(()=>a);
+}