+++ crates/swc_ecma_transforms_compat/tests/block-scoping/issue-5270/3/output.js
@@ -1,0 +1,5 @@
+var _loop = function() {
+    var j = Math.random();
+    setTimeout(()=>console.log(j), 0);
+};
+while(foo())_loop();